move content of edit-tools-shape-ellipse.js to edit-tools-shape.js and delete unnecessary file
