+++ src/fontra/views/editor/editor.js
@@ -51,8 +51,7 @@ import { HandTool } from "./edit-tools-hand.js";
 import { PenTool } from "./edit-tools-pen.js";
 import { PointerTool } from "./edit-tools-pointer.js";
 import { PowerRulerTool } from "./edit-tools-power-ruler.js";
-import { ShapeToolEllipse } from "./edit-tools-shape-ellipse.js";
-import { ShapeToolRect } from "./edit-tools-shape.js";
+import { ShapeToolEllipse, ShapeToolRect } from "./edit-tools-shape.js";
 import { SceneController } from "./scene-controller.js";
 import { MIN_SIDEBAR_WIDTH, Sidebar } from "./sidebar.js";
 import {
